Structural and namespacing changes to sidebars
Deciduous will correct some shortcomings of Thematic's handling of
sidebars. All sidebars will be called using a wrapper function for
get_sidebar (props Justin Tadlock). This will allow two things. One: a
sane filesystem organization for sidebars in a widget-areas
sub-directory. Two: Using get_sidbar allows plugins namely but also
child themes to access the get_sidebar action hook. This was not
possible in Thematic because not all of the widget areas were loaded
using get_sidebar.

And finally I am reducing the number of default sidebars to Primary,
Secondary, and the 3 subsidiaries. I'm thinking 13 areas is a bit much
and I think it leads to feature fatigue and a bloated widgets screen in
the admin. I am open to suggestion on this.