Changes to function name spacing in templates and template hierarchy cascade.
Here I intend to encourage the use of templates over outputting vanilla
html markup via action hooks. So if Deciduous outputs any sizable
amount of html it will be found in a template or template part and not
hidden away in a labyrinth of files using action hooks. I would also
suggest that childthemes follow this pattern when adding markup by
using action hooks to load template files using locate_template or
get_template_part.

This should make the parent and child themes more easily readable and
it will maintain a good casacade for a childtheme to override the
parent in a predictable way that frees of the childtheme's
functions.php from clutter.

The number of templates is being reduced here as well. Thematic had a
fair amount of redundancy in is templates and loop iterations.
Deciduous will refine the templates into a smaller more efficient
package that better leverages the power of template hierarchy.

The templates that are being removed are:

404.php in favor of using an index loop that accounts for when there
are no results (props to the underscores theme)
category.php, tag.php and author.php in favor of simply using the
archive.php most of the code in those three templates was redundant.
attachment.php in favor of single.php

Inside the templates there are changes to name spacing for obvious
reasons. Deciduous will maintain plethora of action hooks for child
themes to use. Although I am reducing the parent's reliance on queueing
its own markup using actions. I am also removing filterable markup in
the parent which seemed to originate in Thematic when it was migrating
from xhtml to html5 and providing legacy support.

Additional Notes:
Template parts will be loaded using: locate_template
Sidebars will be loaded using: deciduous_get_sidebar
Functions that are wrappers for action hooks will be prefixed:
deciduous_do_
Actions will be prefixed: deciduous_a_
Filters will be prefixed: deciduous_f_
Pluggable functions will be prefixed deciduous_p_