Returning header-extenstions.php and footer-extensions.php code into header.php and footer.php
Again with the trend of putting power back in to the templates to
increase legibility.

The contents Thematic's head tag was completely abstracted to
header-extenstions.php. Over time it got messy with added functions for
backwards compatibility as Thematic complied theme review requirements.
And Thematic's migration to an html5 doctype added  filterable markup
in header.php which I see as unnecessary now.

I am putting the contents some of those functions back into the
header.php. This puts references to filters and action directly into
the template. So no more reading though a lengthy file to get to the
complete picture of what is happening or what is possible.

The branding and navigation #access are being moved to template-parts.
This will hopefully replace the flexibility/overridability that was
lost with the removal of the action hook thematic_header. Thematic used
the priority of functions hooked to that action to build the header
element.

The footer.php is losing it's filterable markup. It's sidebars in the
subsidiaries are now called in the proper way to allow the get_sidebar
hook to affect them. And finally a wrapping div id="colophon" was added
to contain the div id="siteinfo". I wanted some action hooks and
flexibility for styling and adding content around the footer text in
it's own div.