Updating helpers.php
Some new additions here:

deciduous_table_wrap to allow for styling the will keep tables from
overflowing the mobile layout. They will scroll instead of overflowing.

deciduous_p_meta_separator will keep the thematic meta separator in
tact but allow for filtering what the separator is. And it helps to
create a more logical class attribute. It is also pluggable so if you
don’t want it or want to roll your own then go for it.